requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9        # multipart file uploads
websockets==12.0               # WebSocket support

# ── Image processing ───────────────────────────────────────────────────────
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80  # headless for Render (no GUI dependencies)
scipy==1.13.0                  # KD-tree for color matching
scikit-image==0.23.2           # Additional image utilities

# ── File handling ──────────────────────────────────────────────────────────
python-magic==0.4.27           # MIME type validation for uploads
aiofiles==23.2.1               # Async file I/O

# ── HTTP client (image fetching) ───────────────────────────────────────────
httpx==0.27.0                  # Async HTTP client for Unsplash + Pixabay

# ── Config & validation ────────────────────────────────────────────────────
pydantic==2.7.1                # Data validation (FastAPI uses this internally)
pydantic-settings==2.2.1       # .env loading via BaseSettings
python-dotenv==1.0.1           # .env file support